# openapi.yaml
openapi: 3.0.3
info:
  title: Micropost API
  description: A simple micropost API with CRUD operations
  version: 1.0.0
  contact:
    name: Micropost API
    email: support@micropost.app

servers:
  - url: http://localhost:3000/api
    description: Development server

paths:
  /microposts:
    get:
      operationId: getMicroposts
      summary: Get all microposts
      description: Retrieve a list of all microposts
      tags:
        - Microposts
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Micropost'
                  count:
                    type: integer
                    example: 2

    post:
      operationId: createMicropost
      summary: Create a new micropost
      description: Create a new micropost with a title
      tags:
        - Microposts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  example: "My new micropost"
                  minLength: 1
      responses:
        '201':
          description: Micropost created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Micropost'
                  message:
                    type: string
                    example: "Micropost created successfully"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /microposts/{id}:
    get:
      operationId: getMicropostById
      summary: Get a micropost by ID
      description: Retrieve a specific micropost by its ID
      tags:
        - Microposts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Micropost ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Micropost'
        '404':
          description: Micropost not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'


components:
  schemas:
    Micropost:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the micropost
          example: 1
        title:
          type: string
          description: Title of the micropost
          example: "My first micropost"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2024-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-02T00:00:00.000Z"
      required:
        - id
        - title
        - createdAt

    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error message"
      required:
        - success
        - message

tags:
  - name: Microposts
    description: Operations related to microposts